requirements.txt
# Agent & LLM
langchain>=1.2
langchain-core>=1.2
langchain-nvidia-ai-endpoints>=1.0
langgraph>=1.0
langgraph-prebuilt>=1.0

# API
fastapi>=0.115
uvicorn[standard]>=0.32

# Database (optional)
supabase>=2.0

# Browser & search
playwright>=1.49
duckduckgo-search>=8.0

# Env
python-dotenv>=1.0
